kernel: sched: Rewrite feec() for better energy efficiency and performance
Google's find_energy_efficient_cpu() implementation is overly specific and
contains several nonsensical heuristics that result in both worse energy
efficiency and performance. Since its heuristics are too specific, it often
doesn't even produce a result, which causes a fallback onto the upstream
scheduler's feec() implementation instead since a CPU must be chosen. This
is especially bad because the upstream feec() implementation produces CPU
choices that conflict with Google's feec(), causing tasks to suboptimally
bounce around between different CPUs, burning energy while hurting
performance.

The need for all of the heuristic spaghetti is indicative that the
underlying algorithm is not robust, since it cannot produce decent results
otherwise. Not to mention, it doesn't fairly assess the energy impact of
all possible CPUs because it disregards CPUs based on faulty heuristics.

Scrap the old feec() entirely and replace it with a better implementation
that always produces a result. This new implementation is observed to yield
better CPU latency and throughput, in addition to significantly reduced
energy. The specifics of the new implementation are described in great
detail in comments.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[EmanuelCN: Adapted to msm-4.19 with sultan's help]