fixup! panic_logstore: always sync after writing one line
It seems the driver can return -EINVAL if the seek head is beyond the
size of the partition. This makes sense, because depending on the
uptime, we can potentially have huge logs, and all we have access to is
a mere 4MiB partition.

This will still return EINVAL, but at least it'll give us... the first
4 megabytes of logs.

TODO: only write the last 2 MiB of logs; it seems the bootloader pastes
its own logs somewhere past the half point. I don't know what's the
best way to do this, however...

Change-Id: Iad053c93d1e2cad52cad6debb113575c4496896b

Signed-off-by: Han Sol Jin <hansol@hansol.ca>